@@ -1,13 +1,24 @@
 {
   "dependencies": {
-    "swing": "^4.0.1"
+    "hammerjs": "^2.0.8",
+    "lodash": "^4.17.11",
+    "raf": "^3.4.1",
+    "rebound": "^0.1.0",
+    "sister": "^3.0.1",
+    "vendor-prefix": "^0.1.0"
   },
   "devDependencies": {
-    "webpack": "^1.14.0",
-    "webpack-dev-server": "^1.16.2"
+    "@babel/core": "^7.2.2",
+    "@babel/cli": "^7.2.3",
+    "@babel/preset-env": "^7.2.3",
+    "@babel/polyfill": "^7.2.5",
+    "webpack": "^4.28.4",
+    "webpack-cli": "^3.2.1",
+    "webpack-dev-server": "^3.1.14",
+    "babel-loader": "^8.0.5"
   },
   "private": true,
   "scripts": {
-    "start": "webpack-dev-server --entry ./index.js"
+    "start": "webpack-dev-server --open --config ./webpack.config.js"
   }
 }
