@@ -1,34 +1,30 @@
 {
-  "author": {
-    "email": "gajus@gajus.com",
-    "name": "Gajus Kuizinas",
-    "url": "http://gajus.com"
-  },
+  "name": "swing",
+  "description": "A swipeable cards interface. The swipe-left/swipe-right for yes/no input. As seen in apps like Jelly and Tinder.",
+  "version": "3.1.2",
+  "author": "andres-brugarolas",
   "dependencies": {
-    "hammerjs": "^2.0.4",
-    "lodash": "^4.6.1",
-    "raf": "^3.1.0",
-    "rebound": "^0.0.13",
-    "sister": "^3.0.0",
+    "hammerjs": "^2.0.8",
+    "lodash": "^4.17.11",
+    "raf": "^3.4.1",
+    "rebound": "^0.1.0",
+    "sister": "^3.0.1",
     "vendor-prefix": "^0.1.0"
   },
-  "description": "A swipeable cards interface. The swipe-left/swipe-right for yes/no input. As seen in apps like Jelly and Tinder.",
   "devDependencies": {
-    "babel": "^6.5.2",
-    "babel-cli": "^6.18.0",
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-register": "^6.18.0",
-    "chai": "^3.2.0",
-    "eslint": "^3.9.1",
-    "eslint-config-canonical": "^6.1.0",
-    "husky": "^0.11.9",
-    "jsdom": "^6.3.0",
-    "jsonfile": "^2.2.1",
-    "mocha": "^3.1.2",
-    "semantic-release": "^4.3.5",
-    "sinon": "^1.16.1",
-    "webpack": "^1.12.1"
+    "@babel/core": "^7.2.2",
+    "@babel/cli": "^7.2.3",
+    "@babel/preset-env": "^7.2.3",
+    "@babel/polyfill": "^7.2.5",
+    "@babel/register": "^7.0.0",
+    "eslint": "^5.12.0",
+    "eslint-config-canonical": "^16.0.0",
+    "husky": "^1.3.1",
+    "mocha": "^5.2.0",
+    "chai": "^4.2.0",
+    "jsdom": "^13.1.0",
+    "sinon": "^7.2.2",
+    "semantic-release": "^15.13.3"
   },
   "keywords": [
     "swipe",
@@ -37,16 +33,18 @@
   ],
   "license": "BSD-3-Clause",
   "main": "./dist/index.js",
-  "name": "swing",
   "repository": {
     "type": "git",
-    "url": "https://github.com/gajus/swing"
+    "url": "https://github.com/brugarolas/swing"
   },
   "scripts": {
     "build": "babel ./src --out-dir ./dist --copy-files",
     "lint": "eslint ./src ./test",
-    "precommit": "npm run lint && npm run test",
-    "test": "mocha --compilers js:babel-register"
+    "test": "mocha --require @babel/register --exit"
   },
-  "version": "3.1.0"
+  "husky": {
+    "hooks": {
+      "pre-commit": "npm run lint && npm run test"
+    }
+  }
 }
