@@ -2,7 +2,7 @@ import Card from './Card';
 import Direction from './Direction';
 import Stack from './Stack';
 
-export {
+export default {
   Card,
   Direction,
   Stack
